Settings layout, sidebar, and user table.
Created a standard layout that can be easily used by other settings page.
This allows for settings pages to have consistent styles and simplifies
the creation process. A sidebar is also created to navigate the different
settings pages (in the future). An updated version of the table is also
created.